app/models/schemas.py
from typing import Optional

from pydantic import BaseModel


class ParsedEvent(BaseModel):
    title: Optional[str] = None
    date: Optional[str] = None
    end_date: Optional[str] = None
    location: Optional[str] = None
    description: Optional[str] = None
