requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-telegram-bot>=21.0
supabase>=2.0.0
google-genai>=1.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
